requirements.txt
aeiou==0.0.20
alias-free-torch==0.0.6
auraloss==0.4.0
descript-audio-codec==1.0.0
einops==0.7.0
einops-exts==0.0.4
ema-pytorch==0.2.3
encodec==0.1.1
huggingface_hub
importlib-resources==5.12.0
k-diffusion==0.1.1
local-attention==1.8.6
pandas==2.0.2
pedalboard==0.9.8
prefigure==0.0.9
pytorch_lightning==2.1.0 
PyWavelets==1.4.1
safetensors
sentencepiece==0.1.99
# s3fs
torch>=2.4.1
torchaudio>=2.4.1
torchvision>=0.19.1
torchmetrics>=0.11.4
tqdm
transformers==4.45.2
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.9.14
wandb==0.15.4
webdataset==0.2.48
x-transformers<1.27.0
nnAudio==0.3.3
mutagen
gradio==4.37.1
gradio_client==1.0.2
gradio_rangeslider==0.0.6
uvicorn==0.30.1
pydantic==2.8.2
pydantic_core==2.20.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastcore==1.5.51
hear21passt
wget
soundfile
torchlibrosa
ftfy
braceexpand
llvmlite
scikit-learn
h5py
seaborn
regex
progressbar
openl3==0.4.2
soxr
